move get_anytext() function to pycsw.core.util
