+++ pycsw/core/repository.py
@@ -416,7 +416,7 @@ def create_custom_sql_functions(connection):
     for function_object in [
         query_spatial,
         update_xpath,
-        get_anytext,
+        util.get_anytext,
         get_geometry_area,
         get_spatial_overlay_rank
     ]:
@@ -426,22 +426,6 @@ def create_custom_sql_functions(connection):
             len(argspec.args),
             function_object
         )
-
-
-def get_anytext(bag):
-    """
-    generate bag of text for free text searches
-    accepts list of words, string of XML, or etree.Element
-    """
-
-    if isinstance(bag, list):  # list of words
-        return ' '.join([_f for _f in bag if _f]).strip()
-    else:  # xml
-        if isinstance(bag, six.binary_type) or isinstance(bag, six.text_type):
-            # serialize to lxml
-            bag = etree.fromstring(bag, PARSER)
-        # get all XML element content
-        return ' '.join([value.strip() for value in bag.xpath('//text()')])
 
 
 def query_spatial(bbox_data_wkt, bbox_input_wkt, predicate, distance):
